* Went back to old baseline system that reduces drift. (however, is also a bit less accurate the rest of the time)

+++ Main/android/app/src/main/java/com/lucidlink/ChartManager.java
@@ -406,7 +406,7 @@ class ChartManager {
 		MainActivity.main.runOnUiThread(() -> {
 			debugText.setText(
 				"1VS2: " + eegProcessor.channel1VSChannel2Strength_averageOfLastX + "\n"
-				+ "EyePos: " + eegProcessor.eyePosX.doubleValue() + "\n"
+				+ "EyePos: " + eegProcessor.eyePosX + "\n"
 				+ "EyePosRel: " + eegProcessor.GetCenterPoint()
 			);
 		});
